pyproject.toml
[tool.poetry]
name = "ai-proxies"
version = "0.1.0"
description = "A collection of proxy servers for various AI services"
authors = ["Clay Hambrick <dclayh@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<4.0"
requests = "^2.31.0"
python-dotenv = "^1.0.0"
msgraph-core = "1.3.*"
msgraph-sdk = "1.26.*"
azure-identity = "^1.15.0"
pandas = "^2.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"